Split CustumerServiceTests to remove presentation test
