+++ abc-bank-facade/BankFacade.cs
@@ -7,11 +7,18 @@ namespace AbcBank.Facade
     public class BankFacade
     {
         private readonly BankPresenter _bankPresenter = new BankPresenter();
+        private readonly StatementPresenter _statementPresenter= new StatementPresenter();
+
         private readonly BankService _bankService = new BankService();
 
         public string CustomerSummary(BankModel bank)
         {
             return _bankPresenter.ToString(_bankService.GetCustomerSummaries(bank));
+        }
+
+        public string GetStatement(CustomerModel customer)
+        {
+            return _statementPresenter.GetStatement(customer);
         }
     }
 }
