+++ abc-bank-tests/abc-bank-tests.csproj
@@ -51,7 +51,8 @@
   </Choose>
   <ItemGroup>
     <Compile Include="PresentationTests\BankPresenterTests.cs" />
-    <Compile Include="CustomerTest.cs" />
+    <Compile Include="ServicesTests\CustomerServiceTests.cs" />
+    <Compile Include="PresentationTests\StatemnetPresenterTests.cs" />
     <Compile Include="Properties\AssemblyInfo.cs" />
     <Compile Include="ServicesTests\BankServiceTests.cs" />
     <Compile Include="TransactionTest.cs" />
